Implement Go to definition in classpath using Ctags module.
The implementation in this commit is very hacky since I'm not sure how
we can jump to definition inside entries of jars. I'm sure it's possible
to just give the jar uri or something and everything automagically
works.

We index the entire classpath for every entry in the project/config
matrix without any deduplication. This needs to be optimized.

Most of the work was to get the compiler to correctly download the
sources jars from the classpath. I had to wrap the call to `Jars.fetch`
inside of a `Future` to unblock the compiler for future updates.
This is a bug that needs to be fixed, we should not be randomly
sprinkling blocks inside `Future  { ... }`.