Implement java ctags with qdox instead of javaparser.
While profiling the javaparser ctags indexer I found out that it
was only indexing around 10k loc/s. The scalameta parser is super slow
and it can index 30-40k loc/s, so this is unacceptable. This commit,
reimplements the java indexer using qdox, which seems to be able to spin
through 250k loc/s, which is a nice improvement :) The qdox api is also
a lot nicer, so this turned out to be a big win!