Properly keep track of argument count in signature help.
Previously, we found the first open paren but now we keep a stack of how
many open parens are in between.
This change makes it possible to return the correct active parameter,
as well as filtering out overloads that have too few parameters.
This also means we no longer return bogus results when outside of
an argument list.