Add special handling for type parameters and varargs.
Previously, signature help didn't work for varargs or applications with
explicit type parameters. This commit generalizes the open/close trick
to work for both brackets and parens, this opens up the possibility
to show signature help for type parameters too in the future.

Varargs are also special cased since the argument count does not match
the length of the method's parameter list.