Pick correct compiler global for new files.
Previously, it was necessary to re-run scalametaEnableCompletions
whenever a new file was created. This commit makes the server pick the
global that is responsible for that source directory. It's still
necessary to lookup by sources as we did before because some builds add
a single file to `sources` without updating
`unmanagedSourceDirectories`.