Add "Run doctor" command to troubleshoot issues.
This commit implements a new server command "Run doctor" that allows the
user to see an overview of all potential problems in the build. The idea
is that whenever users think something is not working they can "Run
doctor" to see what's going on.

Thanks to doctor, we can make the `sbt metalsEnable` more robust by only
modifiying projects with exact supported Scala version.  After build
import, we check with the doctor if there are misconfigured project.

To enable "Run doctor", we add a new Metals LSP extension
`metals/executeClientCommand` which is a clone of
`workspace/executeCommand` except
- it's a notification, not a request
- the server sends the notification to the client, not the other way
  around.
This extension is necessary for the server to be able to display the
doctor page rendered as HTML. In VS Code, we open a webview inside the
editor but for Vim (and other editors) the doctor page is displayed
by the HTTP client.